+++ CHANGELOG.md
@@ -12,6 +12,10 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - API for BiologicalRelationships -`api/v1/biological_relationships?extend[]=biological_property` [#3525]
 - Organization to Attribution slice in Radial annotator [#3529]
 
+### Changed
+
+- Reverted index-based TaxonName autocomplete to comprehensive version
+
 ### Fixed
 
 - Common name language ISO when no language [#3530]